fix adding multiple restaurant users at once; rename userAddition to addUser
